code quality enhacements
